requirements.txt
# Core dependencies
pydantic==2.5.0
python-dotenv==1.0.0
PyPDF2
docx
requests==2.31.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Database
duckdb==0.9.2
psycopg2-binary==2.9.9  # PostgreSQL driver for RDS
pymysql==1.1.0  # MySQL driver for RDS (alternative)

# Testing
pytest==7.4.3
hypothesis==6.92.0
pytest-asyncio==0.21.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Security
bcrypt==4.1.2

# Text processing
textstat==0.7.3

# Web framework
nicegui==1.4.0
fastapi==0.108.0
uvicorn==0.25.0

# Utilities
python-multipart==0.0.6

# AI/ML
huggingface_hub>=1.5.0
